+++ tests/functional/pytorch/jit/016_jit_resnet_torchtrt_aot_backend.py
@@ -12,6 +12,7 @@ from logging import INFO, basicConfig
 
 import timm
 import torch
+from _tuning_data_artifacts import collect_tuning_data
 
 from aitune.torch.backend.torch_tensorrt_aot_backend import TorchTensorRTAotBackend
 from aitune.torch.jit.config import config
@@ -29,6 +30,7 @@ def create_resnet():
     return timm.create_model("resnet18", pretrained=False).to("cuda")
 
 
+@collect_tuning_data(__file__)
 def test_jit_resnet():
     resnet = create_resnet()